web-ui: minor refactor
